+++ mtpy/modeling/ModEM/__init__.py
@@ -5,5 +5,8 @@ from .model import Model
 from .residual import Residual
 from .control_inv import ControlInv
 from .control_fwd import ControlFwd
+from .convariance import Covariance
 
-__all__ = ['ModEMError', 'DataError', 'Stations', 'Data', 'Model', 'Residual', 'ControlInv', 'ControlFwd']
+__all__ = ['ModEMError', 'DataError', 'Stations', 'Data', 'Model', 'Residual', 'ControlInv', 'ControlFwd', 'Covariance']
+
+
